Untested initial changes for general, newer K1C firmwares.

+++ helper.sh
@@ -8,7 +8,7 @@ HELPER_SCRIPT_FOLDER="$(dirname "$(readlink -f "$0")")"
 for script in "${HELPER_SCRIPT_FOLDER}/scripts/"*.sh; do . "${script}"; done
 for script in "${HELPER_SCRIPT_FOLDER}/scripts/menu/"*.sh; do . "${script}"; done
 for script in "${HELPER_SCRIPT_FOLDER}/scripts/menu/K1/"*.sh; do . "${script}"; done
-for script in "${HELPER_SCRIPT_FOLDER}/scripts/menu/K1C_2025/"*.sh; do . "${script}"; done
+for script in "${HELPER_SCRIPT_FOLDER}/scripts/menu/K1C/"*.sh; do . "${script}"; done
 for script in "${HELPER_SCRIPT_FOLDER}/scripts/menu/3V3/"*.sh; do . "${script}"; done
 for script in "${HELPER_SCRIPT_FOLDER}/scripts/menu/3KE/"*.sh; do . "${script}"; done
 for script in "${HELPER_SCRIPT_FOLDER}/scripts/menu/10SE/"*.sh; do . "${script}"; done
@@ -75,7 +75,13 @@ function detect_model() {
   if (command -v get_sn_mac.sh > /dev/null); then
     get_model=$( get_sn_mac.sh model 2>&1 || get_sn_mac.sh model_str 2>&1 )
     if echo "$get_model" | grep -iq "K1C"; then
-      model="K1C_2025"
+      if echo $(get_sn_mac.sh board 2>&1) | grep -iq "CR4SU200382C13"; then
+        model="K1C_2025"
+      elif echo $(get_sn_mac.sh board 2>&1) | grep -iq "CR4CU220812S12"; then
+        model="K1C_X2000E"
+      else
+        echo "Unsupported model!" > /dev/stderr
+      fi
     elif echo "$get_model" | grep -iq "K1"; then
       model="K1"
     elif echo "$get_model" | grep -iq "F001"; then
@@ -104,7 +110,7 @@ if [ ! -L "$BIN_FOLDER"/helper ]; then
 fi
 rm -rf /root/.cache
 
-if [ -z "$model" ] && [ "$model" != "K1C_2025" ] && [ ! -f $INITD_FOLDER/S58factoryreset ]; then
+if [ -z "$model" ] && [[ "$model" != "K1C_2025" ] && [ "$model" != "K1C_X2000E" ]] && [ ! -f $INITD_FOLDER/S58factoryreset ]; then
   cp "$HS_FILES/services/S58factoryreset" $INITD_FOLDER/S58factoryreset
   chmod 755 $INITD_FOLDER/S58factoryreset
 fi
